rest_api_movies/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.2
django-arrayfields==0.1.0
django-cors-headers==3.14.0
django-redis==5.2.0
django-redis-cache==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
